refactor: move hooks to global folder
